move tab request map into ChatCommunicationManager
