+++ plugins/amazonq/shared/jetbrains-community/src/software/aws/toolkits/jetbrains/services/amazonq/lsp/AmazonQLanguageClientImpl.kt
@@ -36,8 +36,6 @@ import software.aws.toolkits.jetbrains.services.codewhisperer.customization.Code
 import software.aws.toolkits.jetbrains.settings.CodeWhispererSettings
 import java.util.UUID
 import java.util.concurrent.CompletableFuture
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.TimeUnit
 
 /**
  * Concrete implementation of [AmazonQLanguageClient] to handle messages sent from server
@@ -116,9 +114,7 @@ class AmazonQLanguageClientImpl(private val project: Project) : AmazonQLanguageC
 
     override fun openTab(params: OpenTabParams): CompletableFuture<OpenTabResult> {
         val requestId = UUID.randomUUID().toString()
-        val result = CompletableFuture<OpenTabResult>()
-
-        pendingTabRequests[requestId] = result
+        val result = ChatCommunicationManager.getInstance(project).addPendingOpenTabRequest(requestId)
 
         val uiMessage = """
                 {
@@ -128,13 +124,6 @@ class AmazonQLanguageClientImpl(private val project: Project) : AmazonQLanguageC
                 }
         """.trimIndent()
         AsyncChatUiListener.notifyPartialMessageUpdate(uiMessage)
-
-        result.orTimeout(30000, TimeUnit.MILLISECONDS)
-            .whenComplete { _, error ->
-                if (error != null) {
-                    pendingTabRequests.remove(requestId)
-                }
-            }
 
         return result
     }
@@ -194,11 +183,5 @@ class AmazonQLanguageClientImpl(private val project: Project) : AmazonQLanguageC
 
     companion object {
         private val LOG = getLogger<AmazonQLanguageClientImpl>()
-
-        private val pendingTabRequests = ConcurrentHashMap<String, CompletableFuture<OpenTabResult>>()
-
-        fun completeTabOpen(requestId: String, tabId: String) {
-            pendingTabRequests.remove(requestId)?.complete(OpenTabResult(tabId))
-        }
     }
 }
